Fix bug where methods weren't implemented in SnakesAndLaddersPlayer
